requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pyTelegramBotAPI==3.7.1
pytz==2020.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
escraper@git+https://git@github.com/ArtemErmulin/escraper.git@29b40ca0e5558c5ebe5d77f05cd7fe4ebd00c18d
notion==0.0.25
